breast_canser.py/requirements.txt
streamlit
scikit-learn
joblib
pandas
numpy
matplotlib
Pillow
